backend/requirements.txt
Flask
Flask-Cors
Flask-Restful
Flask-SQLAlchemy
Flask-Login
Flask-JWT-Extended
redis